feat: screen and business logic for My bids function
