+++ package.json
@@ -62,7 +62,7 @@
     "start": "react-app-rewired start",
     "cap:start:ios": "ionic cap open ios",
     "cap:start:android": "ionic cap open android",
-    "build": "react-app-rewired build",
+    "build": "rimraf ./build && react-app-rewired build",
     "build:android:ci": "npx jetify && npx ionic capacitor sync android && npx ionic capacitor update android && npx ionic capacitor copy android && cd android && ./gradlew assembleDebug && cd ..",
     "test": "react-app-rewired test --transformIgnorePatterns \"node_modules/(?!@ionic)/\"",
     "test:ci": "react-app-rewired test --ci --watchAll=false --transformIgnorePatterns \"node_modules/(?!@ionic)/\"",
