+++ src/components/HomeMarket.tsx
@@ -9,8 +9,8 @@ import { FontSize } from '../constants/font-size';
 import { FontWeight } from '../constants/font-weight';
 import { flagImage } from '../constants/images';
 import AuctionableCardsPage from '../pages/AuctionableCardsPage';
-import AuctionsOwnedPage from '../pages/AuctionsOwnedPage';
-import AuctionParticipateIn from '../pages/AuctionParticipateInPage';
+import AuctionsMyAuctionsPage from '../pages/AuctionsMyAuctionsPage';
+import AuctionParticipateInPage from '../pages/AuctionParticipateInPage';
 
 const ImageBgCol = styled(IonCol)`
   position: relative;
@@ -95,10 +95,10 @@ const HomeMarket: FC = () => {
       <AuctionableCardsPage />
     )}  
     {auctionSellView && (
-      <AuctionsOwnedPage />
+      <AuctionsMyAuctionsPage />
     )}      
     {auctionBuy && (
-      <AuctionParticipateIn/>
+      <AuctionParticipateInPage/>
     )}      
     {!auctionSell && !auctionSellView && !auctionBuy && (
       <IonGrid className="ion-no-padding">
