+++ src/pages/AuctionMyBiddedCardsPage.tsx
@@ -1,289 +1,229 @@
-import { FC, useMemo, useCallback } from 'react';
-import { arrowBackOutline } from 'ionicons/icons';
-import { CalendarOutline } from 'react-ionicons'
-import { Controller, useForm } from 'react-hook-form';
-import { shallowEqual, useSelector } from 'react-redux';
-import { useHistory, useParams } from 'react-router';
+import { FC, useCallback, useContext, useEffect, useMemo, useState } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 import styled from 'styled-components';
-import { Styles } from '../utils/styles';
+import {
+  IonCol,  
+  IonGrid,
+  IonRow,
+  IonSpinner,
+} from '@ionic/react';
+import CardOnAuction from '../components/CardOnAuction';
+import { JSX } from '@ionic/core';
+import Button from '../components/ionic/Button';
+import Label from '../components/ionic/Label';
+import Text from '../components/ionic/Text';
 import { FontSize } from '../constants/font-size';
 import { FontWeight } from '../constants/font-weight';
-import Title from '../components/ionic/Title';
-import Button from '../components/ionic/Button';
-import { JSX } from '@ionic/core';
-
-import {
-  IonList,
-  IonButton,
-  IonItem,
-  IonFooter,
-  IonToolbar,
-  IonInput,
-  IonLabel,
-  IonCol,
-  IonContent,
-  IonGrid,
-  IonPage,
-  IonRow,
-  IonIcon,
-  IonSpinner,  
-} from '@ionic/react';
-import { BaseResourcesTypes } from '@protokol/client';
-import Header from '../components/Header';
-import Img from '../components/ionic/Img';
-import { addIPFSGatewayPrefix } from '../constants/images';
+import useIsMounted from '../hooks/use-is-mounted';
+import { CardTagType } from '../components/CardTag';
+import { AuthLoginContext } from '../providers/AuthLoginProvider';
+import { CollectiblesOnAuctionLoadAction } from '../store/actions/collections';
 import { collectionSelector } from '../store/selectors/collections';
+import { auctionBalloonImage } from '../constants/images';
+import AuctionParticipateInPage from './AuctionParticipateInPage';
 
-const Footer = styled(IonFooter)`
-  position: fixed;
-  bottom: 0;
-`;
+const ImageBgCol = styled(IonCol)`
+  position: relative;
+  padding: 3.5rem 2.05rem 1.85rem 1.5rem;
+  overflow: hidden;
+  z-index: 2;
+  height: 229px !important;
 
-const AmountIonInput = styled(IonInput)<{
-  fontSize?: FontSize;
-}>`
-  ${({ fontSize }) => Styles.serializeFontSize(fontSize)} 
-  font: normal normal bold 21px/28px Open Sans;
-  letter-spacing: 0px;
-  color: #F8C938;
-  opacity: 1;
-`;
+  &:before {
+    content: ' '; 
+    display: block;
+    position: absolute;
+    left: 0;
+    bottom: 0;
+    width: 100%;
+    height: 100%;
+    z-index: 1;
+    opacity: 0.8;
+    background-image: url('${auctionBalloonImage.src}');
+    background-size: cover;    
+    background-repeat: no-repeat;
+    background-position: 50% 0;
+  }
 
-const FooterButton = styled(Button)`
-  &.ion-color-auction {
-    --ion-color-base: var(--ion-color-auction-base);
-    --ion-color-base-rgb: var(--ion-color-warning-rgb);
-    --ion-color-contrast: var(--ion-color-auction-tint);
-    --ion-color-contrast-rgb: var(--ion-color-auction-tint);
-    --ion-color-shade: var(--ion-color-warning-shade);
-    --ion-color-tint: var(--ion-color-auction-tint);
+  > * {
+    position: relative;
+    z-index: 2;
   }
 `;
 
-const AmountIonLabel = styled(IonLabel)`
-  text-align: center;
-  font: italic normal normal 12px/17px Open Sans;
-  letter-spacing: 0px;
-  color: #707070;
-  opacity: 1;
-`;
-
-/*
-const CardIonDatetime = styled(IonDatetime)`
-  text-align: left;
-  font: normal normal bold 14px/19px Open Sans;
-  letter-spacing: 0px;
-  color: #707070;
-  opacity: 1;
-  width: 90%;   
-`;
-*/
-const HorizontalLine = styled.div`
-  width:1px;
-  height:1px;
-  width: 80%;
-  margin-left: 10%;  
-  background:#E6E6E6;
-  position: relative;
-`;
-
-const VerticalLine = styled.div`
-  width:1px;
-  height:75px;
-  background:#E6E6E6;
-  margin-top: 15px;
-  margin-bottom: 15px;
-  position: relative;
-`;
-
-const ViewCardIonButton = styled(Button)<JSX.IonButton>`
+const BiddedCardsButton = styled(Button)<JSX.IonButton>`
   text-decoration: underline;
-  text-transform: capitalize;
-  text-align: left;
-  font: normal normal bold 12px/17px Open Sans;
-  letter-spacing: 0px;
-  color: #F8C938;
-  opacity: 1;
-  border: 0px;
-  box-shadow: none !important;
+  text-transform: uppercase;
+`;
+
+const CollectablesIonRow = styled(IonRow)`
+  overflow: auto;
+  max-height: calc(100vh - 92px - 4.8rem);
 `;
 
 const AuctionMyBiddedCardsPage: FC = () => {
+
+    const { isError, error, isLoading, assets, isLastPage /*, query*/ } = useSelector(collectionSelector);
+
+    const [navToNewBid, setNavToNewBid] = useState(false);
+
+    const navigateNewBid = () => {    
+      setNavToNewBid(true);   
+    }    
+
+    const dispatch = useDispatch();
+    const {
+      session: { publicKey },
+    } = useContext(AuthLoginContext);
   
-  const { handleSubmit, control } = useForm();  
-  const { auctionId, assetId } = useParams<{ auctionId: string, assetId: string }>();  
-  const history = useHistory();
-
-  const submitForm = useCallback( 
-    () => {  
-      //console.log("submitted:" + JSON.stringify(data, null, 2));    
-      history.push(`/market/card/cancelauction/${auctionId}`);
-    },
-    [auctionId, history]
-  );
+ 
+    const isMounted = useIsMounted();
+    useEffect(() => {
+      if (publicKey && isMounted) {
+          dispatch(CollectiblesOnAuctionLoadAction(publicKey!, false, true, undefined));
+      }
+    }, [isMounted, dispatch, publicKey]);
   
-  const { assets } = useSelector(collectionSelector, shallowEqual);
-  const asset = useMemo(() => assets.flat().find(({ id }) => id === assetId), [
-    assets,
-    assetId,
-  ]) as BaseResourcesTypes.Assets; 
+    const loadNextPage = useCallback(() => {
+      if (publicKey) {
+        //const { page } = query ?? { page: 1 };
+        dispatch(CollectiblesOnAuctionLoadAction(publicKey!, false, true, undefined));
+      }
+    }, [/*query,*/ dispatch, publicKey]);
   
-  if (!asset) {
-    return <IonSpinner color="light" />;
-  }
-
-  const {
-    attributes,
-  } = asset;
-
-  const {
-    ipfsHashImageFront,
-    title,
-    subtitle,
-    minimumBid,
-    finalBiddingDate,
-    issuedDate,
-    currentBid    
-  } = attributes as any;
-
-  return (    
-    <IonPage>
-      <Header
-        title={title}
-        buttonTopLeft={
-          <IonButton onClick={() => history.replace('/home')}>
-            <IonIcon color="light" slot="icon-only" icon={arrowBackOutline} />
-          </IonButton>
+    const flatAssets = useMemo(() => assets.flat(), [assets]);
+  
+    const onCardsScroll = useCallback(
+      ({ target }) => {
+        const isBottom =
+          target.scrollHeight - target.scrollTop === target.clientHeight;
+        if (isBottom && !isLastPage && !isLoading) {
+          loadNextPage();
         }
-      />
-      <IonContent fullscreen>        
-        <form onSubmit={handleSubmit(submitForm)}>
-        <IonGrid className="ion-no-padding">
-          <IonRow>
-            <IonCol className="ion-no-padding" size="12">
-              <Img src={addIPFSGatewayPrefix(ipfsHashImageFront)} />
-            </IonCol>
-          </IonRow>
-          <IonRow style={{marginTop:"20px"}}>
-            <IonCol size="5" offset="1">
-              {title && (
-                <Title fontSize={FontSize.L} fontWeight={FontWeight.BOLD} style={{fontFamily:"Open Sans",color:"#252732"}}>
-                  {title}
-                </Title>
-              )}
-              {subtitle && (
-                <Title fontSize={FontSize.SM} style={{fontFamily:"Open Sans",color:"#252732"}}>
-                  {subtitle}
-                </Title>
-              )}
-              {issuedDate && (
-                <Title fontSize={FontSize.SM} style={{fontStyle: 'italic', fontFamily:"Open Sans",color:"#707070"}}>
-                  {new Date(issuedDate).toDateString()}
-                </Title>
-              )}
-              <ViewCardIonButton                 
-                color="#F8C938"                
+      },
+      [loadNextPage, isLastPage, isLoading]
+    );
+      
+  return (
+    <>
+    {navToNewBid && (
+      <AuctionParticipateInPage />
+    )}  
+    {!navToNewBid && (      
+    <>
+     <IonGrid className="ion-no-padding">
+      <IonRow>
+        <ImageBgCol size="12">
+          <Label
+            style={{maxWidth:"256px"}}
+            color="light"
+            fontSize={FontSize.L}
+            fontWeight={FontWeight.BOLD}
+          >
+            Are you still the highest bidder?
+          </Label>
+          <br />
+          <Text
+            style={{maxWidth:"256px"}}
+            className="ion-margin-vertical"
+            color="light"
+            fontSize={FontSize.XS}
+          >
+            Keep an eye on all your bids, and 
+            make sure you don’t loose your cards.
+          </Text>    
+          <br />
+          <IonGrid class="ion-justify-content-center">
+            <IonRow class="ion-justify-content-center">              
+              <IonCol class="ion-align-self-center">
+                <BiddedCardsButton
+                  className="ion-float-right"
+                  fill="clear"
+                  fontSize={FontSize.SM}
+                  color="light"
+                  fontWeight={FontWeight.BOLD}
+                  onClick={navigateNewBid}
+                >
+                  Start a new bid
+                </BiddedCardsButton>                   
+              </IonCol>              
+            </IonRow>
+          </IonGrid>                             
+        </ImageBgCol>            
+      </IonRow>
+    </IonGrid>   
+    <IonGrid className="ion-no-padding">
+      <CollectablesIonRow onScroll={onCardsScroll}>
+          {!isLoading && isError && (
+            <IonCol size="12" class="ion-text-center ion-padding-top">
+              <Text color="danger" fontSize={FontSize.XXL}>
+                Something went wrong!
+              </Text>
+              <Text
+                className="ion-padding-top"
                 fontSize={FontSize.SM}
-                fontWeight={FontWeight.BOLD}
-                onClick={() => history.push(`/home/card/${assetId}`)}>
-                  View Full Card
-              </ViewCardIonButton>               
-            </IonCol>
-            <IonCol size="6" style={{paddingBottom: "60px", textAlign:"center"}}>
-              <Title fontSize={FontSize.XXL} style={{fontFamily:"Open Sans", color:"#F8C938", fontWeight:"bold"}}>
-                  1
-              </Title>
-              <Title fontSize={FontSize.SM} style={{fontStyle: 'italic', fontFamily:"Open Sans",color:"#707070"}}>
-                  Card Available
-              </Title>              
-            </IonCol>
-          </IonRow>
-          <HorizontalLine/>          
-          <IonRow>            
-            <IonCol className="ion-text-center" >              
-              <IonList no-lines lines="none">
-                <IonItem className="ion-text-center" no-lines>     
-                  <Controller
-                    render={() => (
-                      <AmountIonInput
-                        type="text"
-                        color="warning"
-                        name="minimumbid"
-                        value={"$" + minimumBid}
-                        disabled={true}                        
-                      />
-                    )}
-                    control={control}
-                    name="minimumbid"
-                  />                            
-                </IonItem>             
-              </IonList>
-              <AmountIonLabel position="stacked">Minimum Bid</AmountIonLabel> 
-            </IonCol>
-            <VerticalLine/>
-            <IonCol className="ion-text-center" >              
-              <IonList no-lines lines="none">
-                <IonItem className="ion-text-center" no-lines>
-                <Controller
-                    render={() => (
-                      <AmountIonInput
-                        type="text"
-                        color="success"
-                        name="highestbid"                          
-                        value={"$" + currentBid}
-                        disabled={true}                        
-                      />
-                    )}
-                    control={control}
-                    name="highestbid"
-                  />                                                             
-                </IonItem>
-              </IonList>
-              <AmountIonLabel position="stacked">Current Highest Bid</AmountIonLabel> 
-            </IonCol>            
-          </IonRow>
-          <HorizontalLine/>
-          <IonRow style={{marginTop:"10px"}}>
-            <IonCol size="10" offset="1">              
-              <AmountIonLabel position="stacked" style={{marginLeft:"30px"}}>Auction Ends On:</AmountIonLabel> 
-              <IonList no-lines>
-                <IonItem style={{border:"1px solid #707070"}} className="ion-text-center">              
-                  {finalBiddingDate}
-                  { /*
-                  <Controller
-                      as={
-                        <CardIonDatetime value={new Date(finalBiddingDate).toISOString()} className="ion-text-left" displayFormat="MM / DD / YYYY"/>              
-                      }                      
-                      disabled={true}
-                      control={control}
-                      name="finalbiddingdate"      
-                  /> 
-                  */ }                                    
-                  <CalendarOutline color={'#707070'} title="Final Bidding Date"/>
-                </IonItem>                
-              </IonList>
-            </IonCol>
-          </IonRow>             
-        </IonGrid>
-        <Footer className="ion-no-border">
-          <IonToolbar>
-            <FooterButton
-              color="danger"
-              size="large"
-              type="submit"
-              className="ion-text-uppercase ion-no-margin"
-              fontSize={FontSize.SM}
-              fontWeight={FontWeight.BOLD}
-              radius={false}
-              expand="block"
+                color="light"
               >
-              Cancel Auction
-            </FooterButton>
-          </IonToolbar>
-        </Footer>  
-        </form>
-      </IonContent>
-    </IonPage>    
+                {error}
+              </Text>
+            </IonCol>
+          )}
+          {!isLoading && !!flatAssets.length && !isError && (
+            <>
+              {flatAssets.map(({ id, attributes }) => {
+                const {
+                  title,
+                  ipfsHashImageFront,
+                  tags,
+                  timeRemaining,
+                  currentBid,
+                  yourBid                  
+                } = attributes as any;
+
+                const type = 
+                  Array.isArray(tags) && tags.length
+                    ? tags[0]
+                    : CardTagType.None;
+
+                return (
+                  <>
+                    <CardOnAuction
+                      key={id}
+                      id={id}
+                      title={title}
+                      imgIpfsHash={ipfsHashImageFront}
+                      type={type}
+                      linkto={"/market/card/newbid/" + id}  
+                      timeRemaining={timeRemaining}                   
+                      currentBid={currentBid}
+                      yourBid={yourBid}
+                    />
+                  </>
+                );
+              })}
+            </>
+          )}
+          {!isLoading && !flatAssets.length && !isError && (
+            <IonCol size="12" class="ion-text-center">
+              <Text
+                className="ion-padding-top"
+                fontSize={FontSize.L} 
+                color="primary"
+              >
+                No available auctions yet!
+              </Text>
+            </IonCol>
+          )}
+          {isLoading && !flatAssets.length && (
+            <IonCol size="12" class="ion-text-center ion-padding-top">
+              <IonSpinner color="primary" />
+            </IonCol>
+          )}
+        </CollectablesIonRow>      
+    </IonGrid>   
+    </>
+    )} 
+    </>
   );
 };
